+++ tests/test_searches.py
@@ -1,7 +1,10 @@
 import unittest
 
-from allalgorithms.searches import (binary_search, jump_search)
-
+from allalgorithms.searches import (
+    binary_search,
+    fibonacci_search,
+    jump_search
+)
 
 class TestSearches(unittest.TestCase):
 
@@ -12,13 +15,19 @@ class TestSearches(unittest.TestCase):
 		self.assertEqual(None, binary_search(arr, 8))
 		self.assertEqual(None, binary_search(arr, -1))
 
+	def test_fibonacci_search(self):
+		arr = [1, 2, 3, 7, 10, 19, 27, 77]
+		self.assertEqual(3, fibonacci_search(arr, 7))
+		self.assertEqual(7, fibonacci_search(arr, 77))
+		self.assertEqual(None, fibonacci_search(arr, 8))
+		self.assertEqual(None, fibonacci_search(arr, -1))
+    
 	def test_jump_search(self):
 		arr = [1, 2, 3, 7, 10, 19, 27, 77]
 		self.assertEqual(3, binary_search(arr, 7))
 		self.assertEqual(7, binary_search(arr, 77))
 		self.assertEqual(None, binary_search(arr, 8))
 		self.assertEqual(None, binary_search(arr, -1))
-
 
 if __name__ == '__main__':
 	unittest.main()