+++ changelog.md
@@ -8,11 +8,12 @@ Date: October 9, 2018
 
 ### Algorithms:
 
-- ### Sorting
-  - Bubble Sort
 - ### Searches
   - Merge Sort
 
+- ### Sorting
+  - Bubble Sort
+  
 # Version `0.0.1`
 
 Date: TO ADDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD, 2018
@@ -22,10 +23,13 @@ Date: TO ADDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD, 2018
 Added:
 
 - ### Searches
+  - Fibonacci Search
+  - Jump Search
+
+- ### Sorting
   - Bubble Sort
   - Cocktail Shaker Sort
   - Insertion Sort
   - Pigeonhole Sort
   - Selection Sort
   - Stooge Sort
-  - Jump Search